Refactor: Refatorando as request padrões para verificar se quem está acessando é o próprio dono
