+++ app/Http/Requests/BaseRequest.php
@@ -1,0 +1,34 @@
+<?php
+
+namespace App\Http\Requests;
+
+use Illuminate\Foundation\Http\FormRequest;
+
+class BaseRequest extends FormRequest
+{
+    /**
+     * Determine if the user is authorized to make this request.
+     */
+    public function authorize($model): bool
+    {
+        $userId = auth()->id();
+        $modelRequest = $model::find(request()->route('id'));
+        if(!$modelRequest || $modelRequest->user_id !== $userId){
+            throw new \Exception('Unauthorized', 403);
+        }
+
+        return true;
+    }
+
+    /**
+     * Get the validation rules that apply to the request.
+     *
+     * @return array<string, \Illuminate\Contracts\Validation\ValidationRule|array<mixed>|string>
+     */
+    public function rules(): array
+    {
+        return [
+            //
+        ];
+    }
+}
