+++ app/Http/Requests/Product/DestroyRequest.php
@@ -1,0 +1,10 @@
+<?php
+
+namespace App\Http\Requests\Product;
+
+use App\Http\Requests\BaseRequest;
+
+class DestroyRequest extends BaseRequest
+{
+
+}
