+++ app/Http/Requests/User/ShowRequest.php
@@ -1,0 +1,10 @@
+<?php
+
+namespace App\Http\Requests\Product;
+
+use App\Http\Requests\BaseRequest;
+
+class ShowRequest extends BaseRequest
+{
+
+}
