+++ app/Http/Requests/User/StoreRequest.php
@@ -2,16 +2,16 @@
 
 namespace App\Http\Requests\User;
 
-use Illuminate\Foundation\Http\FormRequest;
+use App\Http\Requests\BaseRequest;
 
-class StoreRequest extends FormRequest
+class StoreRequest extends BaseRequest
 {
     /**
      * Determine if the user is authorized to make this request.
      */
-    public function authorize(): bool
+    public function authorize($model = null): bool
     {
-        return false;
+        return true;
     }
 
     /**
